requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==5.1.6
django-allauth==0.61.1
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
idna==3.10
mysqlclient==2.2.7
oauthlib==3.2.2
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
